I have at theo's request split the buffer cache write limiting diff from the disksort diff. Both are necessary together in order to not have io's starving, however they do address two different problems that can cause this to happen. I've incorporated some feedback into this from dlg and matthew already.
As this corrects a really nasty problem, we believe this is critical
for release.

This is the diff to prevent disksort from starving io's by staying
in the first list as more and more io's are inserted in it, instead of
servicing the second. it does this by limiting the number of inserts
that can be done before an inversion.

(also in ~beck/disksort.diff1 on cvs)