Hi,
As it was recently discovered some newer machines have 40 bit
physical address size which means that the math that is good
for 36 bit paddr just doesn't cut it anymore.  In fact the
memory region lengths we program into the MSRs are completely
off.

The following diff has some code ported from FreeBSD to address
this and other concerns.  It substantiates a comment about TLB
invalidation by doing an actual TLB flush before we enable the
MTRR and also fixes an incorrect behavior and a race spotted by
Theo when we're setting MSR_MTRRdefType to be the same as in
boot CPU.

Please give this diff a spin as it might solve some weird problems
in X as well as fixing the speed problem for me which we believe is
caused by screwing up the uncacheable PCI device memory mappings.